+++ CHANGELOG.md
@@ -7,9 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.39] - 2026-02-??
 
+### :toolbox: Added
+
+- Add a cmdlet to the DiaConvertImageToPDF .Net Assembly (New-ConvertPngToPdf)
+- Add a cmdlet to the Diagrammer .Net Assembly:
+  - New-ConvertPngToPdf
+  - New-WatermarkToImage
+  - Get-ImageHeightFromBase64
+  - Get-ImageHeightFromFile
+  - Get-ImagePropertiesFromFile
+  - Get-ImageWidthFromBase64
+  - Get-ImageWidthFromFile
+  - Get-ResizeImageFromFile
+  - Get-RotateImageFromFile
+
 ### :arrows_clockwise: Changed
 
 - Upgrade assembly libraries
+- Update module version to v0.2.39
 
 ## [0.2.38] - 2026-02-21